// frontend/src/types/api.ts
export interface LibraryPlaylistSummary {
    spotify_id: string;
    name: string;
    owner_name: string;
    owner_is_self: boolean;
    track_count: number;
    images: string;
    public: boolean;
    collaborative: boolean;
}
export interface LibraryAlbumSummary {
    spotify_id: string;
    name: string;
    artists: string;
    images: string;
    release_date: string;
    total_tracks: number;
}